Add Account route and move some table components to UI folder
