+++ .env.development.local
@@ -1,3 +1,2 @@
-FAST_REFRESH=true
 SKIP_PREFLIGHT_CHECK=true
 GENERATE_SOURCEMAP=false
