+++ src/components/Pool/PairView/UserLiquidity/UserLiquidity.tsx
@@ -8,12 +8,12 @@ import styled from 'styled-components'
 import { useTotalSupply } from '../../../../data/TotalSupply'
 import { useActiveWeb3React } from '../../../../hooks'
 import { useTokenBalance } from '../../../../state/wallet/hooks'
+import { DimBlurBgBox } from '../../../../ui/DimBlurBgBox'
 import { getAccountAnalyticsLink } from '../../../../utils'
 import { currencyId } from '../../../../utils/currencyId'
 import { unwrappedToken } from '../../../../utils/wrappedCurrency'
 import { ButtonExternalLink, ButtonPurpleDim } from '../../../Button'
 import { CurrencyLogo } from '../../../CurrencyLogo'
-import { DimBlurBgBox } from '../../DimBlurBgBox/styleds'
 import { InfoGrid } from '../InfoGrid/InfoGrid.styles'
 import { ValueWithLabel } from '../ValueWithLabel'
