+++ src/pages/Account/index.tsx
@@ -1,0 +1,1 @@
+export { Account as default } from './Account'
