+++ src/pages/Pools/LiquidityMiningCampaign/index.tsx
@@ -19,10 +19,10 @@ import { useLiquidityMiningCampaign } from '../../../hooks/useLiquidityMiningCam
 import { useRouter } from '../../../hooks/useRouter'
 import { useTokenBalance } from '../../../state/wallet/hooks'
 import { TYPE } from '../../../theme'
+import { PageWrapper } from '../../../ui/PageWrapper'
 import { currencyId } from '../../../utils/currencyId'
 import { unwrappedToken } from '../../../utils/wrappedCurrency'
 import { ResponsiveButtonPrimary, ResponsiveButtonSecondary } from '../../LiquidityMining/styleds'
-import { PageWrapper } from '../../PageWrapper'
 
 const TitleRow = styled(RowBetween)`
   ${({ theme }) => theme.mediaWidth.upToSmall`
