+++ src/ui/Header.tsx
@@ -1,0 +1,6 @@
+import { Flex } from 'rebass'
+import styled from 'styled-components'
+
+export const Header = styled(Flex)`
+  border-bottom: 1px solid ${({ theme }) => theme.bg3};
+`
