+++ src/ui/HeaderText.tsx
@@ -1,0 +1,9 @@
+import { Text } from 'rebass'
+import styled from 'styled-components'
+
+export const HeaderText = styled(Text)`
+  font-weight: 600;
+  font-size: 10px;
+  color: ${({ theme }) => theme.purple3};
+  text-transform: uppercase;
+`
